+++ docs/api/changelog.rst
@@ -20,7 +20,8 @@ Added
 - Added :meth:`imod.msw.MetaSwapModel.split` to split MetaSWAP models.
 - Added :meth:`imod.mf6.HorizontalFlowBarrierResistance.from_imod5_data` to load
   barriers from 3D GEN files. 
-
+- Added ``name`` argument to:meth:`imod.mf6.Modflow6Simulation.from_imod5_data`
+  to provide custom name to imported simulation and model.
 
 Fixed
 ~~~~~